+++ src/Properties/UntypedProperty.php
@@ -56,4 +56,12 @@ class UntypedProperty implements Property
     {
         return [];
     }
+
+    /**
+     * @inheritdoc
+     */
+    public function setterComment()
+    {
+        return '';
+    }
 }